chore: offer is page, page is offer
